lbmap: Move to pkg/loadbalancer/legacy/lbmap
Relocate the legacy implementation to pkg/loadbalancer/legacy/lbmap.
