+++ cilium-dbg/cmd/bpf_lb_maglev_get.go
@@ -14,8 +14,8 @@ import (
 	"github.com/cilium/cilium/pkg/command"
 	"github.com/cilium/cilium/pkg/common"
 	"github.com/cilium/cilium/pkg/ebpf"
+	"github.com/cilium/cilium/pkg/loadbalancer/legacy/lbmap"
 	"github.com/cilium/cilium/pkg/logging"
-	"github.com/cilium/cilium/pkg/maps/lbmap"
 )
 
 var bpfMaglevGetCmd = &cobra.Command{
